[Delete] 마이페이지 관련 뷰 의존성 주입 / 통신 코드 삭제
